fix: move the game-component into the content-field, correct a game-component logic
