+++ match-match-game/src/components/content-field/game/cards-field/card/card.scss
@@ -39,7 +39,7 @@
   }
   &__back {
     background-color: gray;
-    background-image: url('../../../../assets/card-back.png');
+    background-image: url('../../../../../assets/card-back.png');
     background-position: center;
     background-size: cover;
     transform: rotateY(180deg);
