+++ match-match-game/src/components/content-field/game/cards-field/card/card.ts
@@ -1,5 +1,5 @@
 import './card.scss';
-import { BaseComponent } from '../../../base-component';
+import { BaseComponent } from '../../../../base-component';
 
 const FLIP_CLASS = 'flipped';
 export class Card extends BaseComponent {
